Rename factory classes into parsers
