+++ src/main/java/lawrence/database/TaskFileManager.java
@@ -1,7 +1,7 @@
 package lawrence.database;
 
-import lawrence.factory.InputSource;
-import lawrence.factory.TaskFactory;
+import lawrence.parser.InputSource;
+import lawrence.parser.TaskParser;
 import lawrence.task.Task;
 
 import java.io.File;
@@ -42,7 +42,7 @@ public class TaskFileManager {
         while(sc.hasNext()) {
             try {
                 String taskString = sc.nextLine();
-                tasks.add(TaskFactory.createTask(taskString, InputSource.FILE));
+                tasks.add(TaskParser.createTask(taskString, InputSource.FILE));
             } catch (IllegalArgumentException | DateTimeParseException e) {
                 // Task contains faulty information in file, just skip
             }
