+++ src/main/java/lawrence/parser/InputSource.java
@@ -1,4 +1,4 @@
-package lawrence.factory;
+package lawrence.parser;
 
 public enum InputSource {
     FILE,
