+++ src/main/java/lawrence/parser/TaskCreator.java
@@ -1,4 +1,4 @@
-package lawrence.factory;
+package lawrence.parser;
 
 import lawrence.task.Task;
